+++ tools/pytorch/run.py
@@ -5,6 +5,7 @@ from importlib import import_module
 
 import torch
 
+# Works with JSON with proper input, currently have set to ignore input so doesn't mess with actions
 
 def resolve():
     functions = import_module("pytorchGMM")
